Added hashing logic Initial logic to call the hash function.

+++ app/src/main/java/com/sharadjoshi/mine/bitcoin/data/BlockHeader.kt
@@ -7,6 +7,6 @@ data class BlockHeader(val version : Int = 0) {
     val prevBlockhash: String = ""
     val merkleRoot: String = ""
     val timestamp: Long = 0
-    val difficultyTarget: Int = 0
+    val difficultyTarget: Long = 0
     val nonce: Long = 0
 }
